Minor fixes
Remove start/busy flag from documentation for floats

Improve the logic for hitting the bunkers in spaceinvaders. Allow bombs when dropped to be slightly deflected from centre of invader, gives a better pattern when hitting the bunkers.

gpu remove the 15 - and 7 - when calculating the x location within the tile for blitting, by switching the forward and reverse coordinates.

tilemap correct 0-7 to 0-15 in x position comment
sprites, remove excess space.