Update FPU from Silice-Playground
Use a single set of classifiers for the fpu registers (in CPU), passed to FPU units.

For FPU calculations, use a single normaliser for addsub/divide, and a single rounder and combiner for addsub, mul, div and sqrt.

Use the always loop in FPUcalc to multiplex the outputs from the units to the normaliser/rounder/combiner. Remove same from units.

PAWS FPU no longer uses the float32 from common, uses a modified version that relies on the FPU calc block for some inputs/outputs.